requirements.txt
Flask
Flask-RESTful
Flask-Cors
swagger-ui-py==21.12.8
pymongo
dnspython
requests
python-dotenv==0.17.1
pytest
python-decouple==3.4
gunicorn==20.1.0
cryptography
PyJWT
email_validator
validate_email
python-dateutil
bcrypt
azure-communication-email